+++ crates/teloxide-core/src/payloads/codegen.rs
@@ -64,6 +64,8 @@ fn codegen_payloads() {
                 "SendPaidMedia"
                     | "SendMediaGroup"
                     | "SetBusinessAccountProfilePhoto"
+                    | "PostStory"
+                    | "EditStory"
                     | "EditMessageMedia"
                     | "EditMessageMediaInline"
                     | "CreateNewStickerSet"